Fix build problems

+++ src/realm/array_timestamp.hpp
@@ -61,7 +61,7 @@ public:
     Timestamp get(size_t ndx) const
     {
         util::Optional<int64_t> seconds = m_seconds.get(ndx);
-        return seconds ? Timestamp(*seconds, m_nanoseconds.get(ndx)) : Timestamp{};
+        return seconds ? Timestamp(*seconds, int32_t(m_nanoseconds.get(ndx))) : Timestamp{};
     }
     bool is_null(size_t ndx) const
     {
